+++ sbepistats/tests/bevy_auto_plugin.rs
@@ -4,7 +4,7 @@ use sbepistats::*;
 
 fn app() -> App {
     let mut app = App::new();
-    app.add_plugins((MinimalPlugins, StatsPlugin, TestPlugin));
+    app.add_plugins((MinimalPlugins, TestPlugin));
     app
 }
 
@@ -14,27 +14,63 @@ struct TestPlugin;
 
 #[derive(StatType)]
 #[auto_plugin_build_hook(plugin = TestPlugin, hook = StatTypeHook)]
+#[auto_plugin_build_hook(plugin = TestPlugin, hook = ConfigureStatTypeAddHook)]
 struct Speed;
 
 #[derive(Component)]
-#[auto_plugin_build_hook(plugin = TestPlugin, hook = StatModifierAddHook::<Speed>::default())]
-struct SpeedBoost;
+struct Accelerate;
 
-impl StatModifierAdd<Speed> for SpeedBoost {
+#[auto_system(plugin = TestPlugin, schedule = PreUpdate, config(
+    in_set = StatSystems::<Speed>::Op(DataTypeOp::Add),
+))]
+fn accelerate(mut query: Query<(&mut Stat<Speed>, &Stat<Acceleration>), With<Accelerate>>) {
+    for (mut stat, boost) in query.iter_mut() {
+        stat.add_modifier(boost.total());
+    }
+}
+
+#[derive(StatType)]
+#[auto_plugin_build_hook(plugin = TestPlugin, hook = StatTypeHook)]
+#[auto_plugin_build_hook(plugin = TestPlugin, hook = OrderStatBeforeHook::<Speed>::default())]
+struct Acceleration;
+
+#[derive(Component)]
+#[auto_plugin_build_hook(plugin = TestPlugin, hook = StatModifierAddHook::<Acceleration>::default())]
+struct AccelerationBoost;
+
+impl StatModifierAdd<Acceleration> for AccelerationBoost {
     fn add(&self) -> f32 {
         0.2
+    }
+}
+
+#[derive(StatType)]
+#[auto_plugin_build_hook(plugin = TestPlugin, hook = StatTypeHook)]
+#[auto_plugin_build_hook(plugin = TestPlugin, hook = OrderStatAfterHook::<Speed>::default())]
+#[auto_plugin_build_hook(plugin = TestPlugin, hook = ConfigureStatTypeAddHook)]
+struct Position;
+
+#[derive(Component)]
+struct Move;
+
+#[auto_system(plugin = TestPlugin, schedule = PreUpdate, config(
+    in_set = StatSystems::<Position>::Op(DataTypeOp::Add),
+))]
+fn r#move(mut query: Query<(&mut Stat<Position>, &Stat<Speed>), With<Move>>) {
+    for (mut stat, boost) in query.iter_mut() {
+        stat.add_modifier(boost.total());
     }
 }
 
 #[test]
 fn auto_plugin_f32_stat_without_modifier() {
     let mut app = app();
-    app.world_mut().spawn(Stat::<Speed>::new(1.0));
+    app.world_mut().spawn(Stat::<Acceleration>::new(1.0));
     app.update();
     assert_eq!(
         1.0,
         app.world_mut()
-            .query::<&Stat::<Speed>>()
+            .query::<&Stat::<Acceleration>>()
             .single(app.world())
             .unwrap()
             .total()
@@ -44,12 +80,55 @@ fn auto_plugin_f32_stat_without_modifier() {
 #[test]
 fn auto_plugin_f32_stat_with_modifier() {
     let mut app = app();
-    app.world_mut().spawn((Stat::<Speed>::new(1.0), SpeedBoost));
+    app.world_mut()
+        .spawn((Stat::<Acceleration>::new(1.0), AccelerationBoost));
     app.update();
     assert_eq!(
         1.2,
         app.world_mut()
+            .query::<&Stat::<Acceleration>>()
+            .single(app.world())
+            .unwrap()
+            .total()
+    )
+}
+
+#[test]
+fn auto_plugin_order_stat_before() {
+    let mut app = app();
+    app.world_mut().spawn((
+        Stat::<Speed>::new(1.0),
+        Accelerate,
+        Stat::<Acceleration>::new(1.0),
+        AccelerationBoost,
+    ));
+    app.update();
+    assert_eq!(
+        2.2,
+        app.world_mut()
             .query::<&Stat::<Speed>>()
+            .single(app.world())
+            .unwrap()
+            .total()
+    )
+}
+
+#[test]
+fn auto_plugin_order_stat_after() {
+    let mut app = app();
+    app.world_mut().spawn((
+        Stat::<Position>::new(1.0),
+        Move,
+        Stat::<Speed>::new(1.0),
+        Accelerate,
+        Stat::<Acceleration>::new(1.0),
+        AccelerationBoost,
+    ));
+    app.update();
+    assert_eq!(
+        3.2,
+        app.world_mut()
+            .query::<&Stat::<Position>>()
             .single(app.world())
             .unwrap()
             .total()