+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 0.2.0
+
+Added:
+- The order of stat calculation can be set using `AppExt::order_stats`.
+- The running operator total can be modified with `Stat::add_modifier` in `StatSystems::Op`.
+
+Changed:
+- Removed StatsPlugin.
+
 ## 0.1.0
 
 Initial release!